Переносит логику работы формы в модуль формы
